+++ app/src/main/java/org/evcs/android/model/user/Wrappers.kt
@@ -12,4 +12,4 @@ class NameWrapper(var firstName: String, var lastName: String)
 
 class ChangePasswordWrapper(var email: String, var identifier: String, var password: String, var passwordConfirmation: String)
 
-class BooleanWrapper(var enabled: Boolean)
+class BooleanWrapper(var enabled: String)
